Delete QuickStart Page/ Add Simulation page/ Finish document function/ Add dependent package
Delete QuickStart Page or any example block that need to be done after publish the julia package.
Finalize most of the function in the Light Package. Except some of the function transfer!.
Adding simulation page to write the @docs block.
Adding another dependent package in the hope that it would help documenting the sturcture param